fix(admin): 채팅방 상세 훅 분리
